requirements.txt
Markdown
dependency_injector >= 4.36.0, < 4.37
django >= 3.2.7, < 3.3
django-cors-headers >= 3.8.0, < 3.9
django-filter >= 2.4.0, < 2.5
django-shortuuidfield >= 0.1.3, < 0.2
djangorestframework >=3.12.4, < 3.13
enum34 >=1.1.10, < 1.2
jsonfield >= 3.1.0, < 3.2
numpy >= 1.22, < 1.23
paho-mqtt >= 1.5.1, < 1.6
shortuuid >= 1.0.1, < 1.1
six >= 1.16.0, < 1.17
gunicorn == 20.1.0
smbus2
